Create q710
In every programming language, even the simplest, there is a branching operator that allows you to check the truth of a logical expression and, depending on its result, perform an action. The condition of the branch operator is a Boolean expression that can result in either true or false. Variables that can participate in a Boolean expression are called Boolean. Boolean variables can be combined into complex conditions using logical operations (functions):

AND(x1,x2,.. xs). Operation " And " returns true value if all its operands are true. (2 ≤ number of operands ≤ s);
OR(x1,x2,.. xs). Operation " OR " returns true value if at least one of its operands is true. (2 ≤ number of operands ≤ s);
NOT(x1). Operation "NOT" changes the value of operand x1 to the opposite (operand is always the same).
In a logical condition, multiple logical functions nested within each other can be used, that is, the result of one function can be used by another as an operand. For example, AND(A, B, OR(C,D)). This expression will be true when a, B and (C or D) are true.

You want to write a program that will determine the result of the expression based on the existing logical expression and the value of logical variables. The number of operands of and and OR functions can be any (2 ≤ s ≤ 26).